backend/pyproject.toml
[tool.poetry]
name = "groomin"
version = "0.0.1"
description = "Online drawing game"
authors = ["Foucauld Degeorges <foucauld.degeorges@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.4"
argon2-cffi = "20.1.0"
asgiref = "==3.4.1"
dj-database-url = "0.5.0"
django = "==3.2.6"
django-memoize = "2.3.1"
djangorestframework = "==3.12.4"
djangorestframework_simplejwt = "==4.6.0"
psycopg2-binary = "==2.9.1"
pytz = "==2021.1"
requests = "==2.26.0"
sentry-sdk = "==1.3.1"
sqlparse = "0.4.1"
django-eventstream = "==4.2.0"
django-cors-headers = "==3.7.0"

[tool.poetry.dev-dependencies]
debugpy = "==1.4.1"
flake8 = "==3.9.2"
factory-boy = "==3.2.0"
coverage = "==5.5"
django-debug-toolbar = "==3.2.1"
django-extensions = "==3.1.3"
isort = "==5.9.3"
black = "^19.10b0"

[tool.black]
exclude = 'migrations'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
